+++ NEWS.md
@@ -1,6 +1,7 @@
 ### crunch 1.9.7 (under development)
 * Better support for deleting hidden variables
 * Allow subsetting of datasets to include hidden variables
+* Require that version names must be a single string value
 
 ### crunch 1.9.6
 * Add warning that the `pattern` argument for functions including `makeArray`, `makeMR`, `deleteVariables`, and `hideVariables` is being deprecated. The help pages for those functions advise you to grep for or otherwise identify your variables outside of these functions.